Run unit tests as part of CI (#68)
* Refactor unit tests to avoid displaying plots during testing

* Move unit tests to new directory

* Run unit tests in CI

* fix formatting
